refactor(vfs): move dbOperations from useFileSystem into utils/indexedDB
Decouple the generic IndexedDB CRUD object (getAll/get/put/delete/clear) from
the Finder hook. It now lives alongside ensureIndexedDBInitialized/STORES in
src/utils/indexedDB.ts, and all 13 cross-app consumers (Finder, TextEdit, Paint,
Chats, Terminal, Applet Viewer, Desktop, iPod cache, migration) import it from
there. Pure code move (byte-identical CRUD); useFileSystem imports it back for
internal use. No re-export, so tsc enforces full migration.

Verified end-to-end with Playwright (clean Chromium): TextEdit File>Save writes a
doc to the 'documents' IndexedDB store and it persists across a full page reload
(PERSISTED_AFTER_RELOAD=true). Behavior identical to main.
