+++ src/apps/terminal/commands/open.ts
@@ -1,11 +1,8 @@
 import { Command, CommandContext, CommandResult } from "../types";
 import { appIds, AppId } from "@/config/appIds";
 import { useFilesStore } from "@/stores/useFilesStore";
-import {
-  dbOperations,
-  DocumentContent,
-} from "@/apps/finder/hooks/useFileSystem";
-import { STORES } from "@/utils/indexedDB";
+import { DocumentContent } from "@/apps/finder/hooks/useFileSystem";
+import { STORES, dbOperations } from "@/utils/indexedDB";
 import i18n from "@/lib/i18n";
 
 // App name aliases for convenience
