+++ src/apps/terminal/commands/vim.ts
@@ -1,11 +1,8 @@
 import { Command, CommandContext, CommandResult } from "../types";
 import { useTerminalStore } from "@/stores/useTerminalStore";
 import { useFilesStore } from "@/stores/useFilesStore";
-import {
-  dbOperations,
-  DocumentContent,
-} from "@/apps/finder/hooks/useFileSystem";
-import { STORES } from "@/utils/indexedDB";
+import { DocumentContent } from "@/apps/finder/hooks/useFileSystem";
+import { STORES, dbOperations } from "@/utils/indexedDB";
 
 export const vimCommand: Command = {
   name: "vim",
