+++ src/apps/textedit/hooks/useFileOperations.ts
@@ -15,11 +15,8 @@ import {
   parseRichMarkdown,
   serializeRichMarkdown,
 } from "../utils/richMarkdown";
-import {
-  dbOperations,
-  DocumentContent,
-} from "@/apps/finder/hooks/useFileSystem";
-import { STORES } from "@/utils/indexedDB";
+import { DocumentContent } from "@/apps/finder/hooks/useFileSystem";
+import { STORES, dbOperations } from "@/utils/indexedDB";
 import { getStoreForFile } from "@/utils/indexedDBOperations";
 import { TEXTEDIT_ANALYTICS, track } from "@/utils/analytics";
